Fix jQuery page footer
The jQuery page footer looks off. Why? Because the doc pages are generated with `dist/pre.html` prepended, and `dist/post.html` appended. `pre.html` contains an open div tag for `<div class="content">` that is supposed to be closed in `post.html`. This relies on the page content has correct HTML tag matching, which unfortunately is off by one tag in this page.